+++ src/document/jsonl.rs
@@ -48,6 +48,7 @@
 //! ...
 //! ```
 use crate::document::Document;
+use crate::helper::string::DisplayOnlyForDebugging;
 use crate::DataResult;
 use crate::{DataSet, Metadata};
 use json_value_search::Search;
@@ -124,7 +125,7 @@ impl Document for Jsonl {
                         Some(Value::Array(records)) => {
                             for record in records {
                                 trace!(
-                                    record = format!("{:?}", record).as_str(),
+                                    record = record.display_only_for_debugging(),
                                     "Record deserialized"
                                 );
                                 dataset.push(DataResult::Ok(record));
@@ -132,7 +133,7 @@ impl Document for Jsonl {
                         }
                         Some(record) => {
                             trace!(
-                                record = format!("{:?}", record).as_str(),
+                                record = record.display_only_for_debugging(),
                                 "Record deserialized"
                             );
                             dataset.push(DataResult::Ok(record));
@@ -140,7 +141,7 @@ impl Document for Jsonl {
                         None => {
                             warn!(
                                 entry_path = format!("{:?}", entry_path).as_str(),
-                                record = format!("{:?}", record).as_str(),
+                                record = record.display_only_for_debugging(),
                                 "Entry path not found"
                             );
                             dataset.push(DataResult::Err((
@@ -156,7 +157,7 @@ impl Document for Jsonl {
                 (Ok(Value::Array(records)), None) => {
                     for record in records {
                         trace!(
-                            record = format!("{:?}", record).as_str(),
+                            record = record.display_only_for_debugging(),
                             "Record deserialized"
                         );
                         dataset.push(DataResult::Ok(record.clone()));
@@ -164,7 +165,7 @@ impl Document for Jsonl {
                 }
                 (Ok(record), None) => {
                     trace!(
-                        record = format!("{:?}", record).as_str(),
+                        record = record.display_only_for_debugging(),
                         "Record deserialized"
                     );
                     dataset.push(DataResult::Ok(record.clone()));
@@ -225,7 +226,7 @@ impl Document for Jsonl {
                 };
 
                 trace!(
-                    record = format!("{:?}", value).as_str(),
+                    record = value.display_only_for_debugging(),
                     "Record serialized"
                 );