+++ src/updater/mod.rs
@@ -2,9 +2,9 @@ mod tera;
 pub mod tera_helpers;
 
 use self::tera::Tera;
+use async_trait::async_trait;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::collections::HashMap;
 use std::io;
 
 pub const INPUT_FIELD_KEY: &str = "input";
@@ -44,13 +44,14 @@ impl UpdaterType {
 }
 
 /// Trait to format a field of an object with a template engine and a template field.
+#[async_trait]
 pub trait Updater: Send + Sync {
     /// Update the object with some mapping
-    fn update(
+    async fn update(
         &self,
         object: &Value,
         context: &Value,
-        mapping: &HashMap<String, Vec<Value>>,
+        mapping: &Value,
         actions: &[Action],
     ) -> io::Result<Value>;
 }